feat(router): retire `router status` for the verbs that already answer it
`router status` bundled three unrelated probes, and two of them had been
answered elsewhere for a while. Where Router lives is what every verb resolves
before doing anything, and the one that just failed already said whether it was
missing or invisible to this profile. Who you are is `olares-cli profile
whoami`, which is the reason `router whoami` went in 141b6d5d7 — Router reads
the Olares role rather than keeping one of its own, so there is nothing here
that could disagree with it.

The third probe, Router's own /healthz, goes with it. It is real and has no
replacement, but a verb that exists to say "the thing you were about to use is
up" is answered by using it: an unreachable Router is reported by whichever
verb reached for it, naming the entrance it tried.

The ALLOWED MODELS row this verb grew in 141b6d5d7 is the reason to do it now
rather than later. It printed users.allowed_models, which Router never writes
and never reads for a decision — a permission control that controls nothing.
Restricting what a caller may reach is a property of a key, and a key carrying
an allowlist already sees only those models in `router models` and in
GET /v1/models. The Router side of that field is being taken off the wire in
the same sweep.

identity.go loses fetchConsoleUser and isAdmin, which had no other caller, and
what survives — consoleUser and userLabels — joins users_lookup.go, since both
files were the same subject: turning a name into the id the routes take.
consoleUser shrinks to the three fields those lookups read.

Verification: go build ./..., go vet ./cmd/..., go test ./cmd/ctl/... all
green; gofmt clean on the touched files. Against a live Router: `router --help`
no longer lists status, `key list` still labels its owner through userLabels,
and `usage summary --user nobody-here` still refuses with the names Router
knows.

Out (deferred): `olares-cli router status` prints the group help and exits 0
rather than reporting an unknown command, matching every other group root in
this CLI (`market zzz` behaves the same). Adding an Args validator is a
tree-wide change, not this commit's.
