+++ cli/cmd/ctl/router/users_lookup.go
@@ -12,9 +12,20 @@ import (
 // the first time somebody's session reaches Router, created from the X-BFL-USER
 // header the edge injects, and the role on it is the Olares role read back
 // rather than a second one to keep track of. So the people on this Olares are
-// `olares-cli settings users list`, and what is left here is the lookup the
-// verbs that take a person need: `key issue --for-user`, `quota set --user`,
-// `usage --user` and `audit --user` are given a name and have to send an id.
+// `olares-cli settings users list`, their role is `olares-cli profile whoami`,
+// and what is left here is the lookup the verbs that take a person need: `key
+// issue --for-user`, `quota set --user`, `usage --user` and `audit --user` are
+// given a name and have to send an id.
+
+// consoleUser is the part of a user row those lookups read. The row carries
+// more — a role, a source, a status that the directory's delete cascade
+// tombstones — and none of it is read here, because nothing left in this tree
+// prints a person.
+type consoleUser struct {
+	ID       string `json:"id"`
+	BflName  string `json:"bfl_name"`
+	OlaresID string `json:"olares_id,omitempty"`
+}
 
 func listUsers(ctx context.Context, pc *preparedClient) ([]consoleUser, error) {
 	return collection[consoleUser](ctx, pc, epUsers)
@@ -49,4 +60,23 @@ func resolveUserID(ctx context.Context, pc *preparedClient, ref string) (string,
 		none:  "Router has no user records yet, which happens before anybody has opened it",
 		note:  "A user appears only after their first visit to Router",
 	}.err()
+}
+
+// userLabels maps Router's user ids to the names a person recognises, for the
+// columns that would otherwise print a UUID. Failing to read the list is not
+// worth failing the caller over — the id is still printed — so this returns an
+// empty map rather than an error.
+func userLabels(ctx context.Context, pc *preparedClient) map[string]string {
+	users, err := listUsers(ctx, pc)
+	if err != nil {
+		return map[string]string{}
+	}
+	out := make(map[string]string, len(users))
+	for i := range users {
+		u := &users[i]
+		if u.ID != "" && u.BflName != "" {
+			out[u.ID] = u.BflName
+		}
+	}
+	return out
 }
